Module kernel::mem

Re-exports

pub use address::FrameNumber;

pub use address::PageNumber;

pub use address::PhysicalAddress;

pub use address::VirtualAddress;

pub use frame_allocator::deallocate_frame;

pub use segment::MapPermission;

pub use segment::PageSet;

pub use segment::KERNEL_SPACE;

pub use segment::KERNEL_SPACE;

pub use user_ptr::UserPtr;

Modules

address 🔒

The address module defines various structs for the Sv39 page table specification.

frame_allocator 🔒

The frame_allocator module provides a frame allocator for the kernel.

heap_allocator 🔒

The heap_allocator module provides a heap allocator for the kernel. The heap is initialized with a fixed size as the KERNEL_HEAP_SIZE constant.

page_table 🔒

The page_table module defines a 3-level page table that follows the RISC-V Sv39 page table specification. The page table supports 512 GB of virtual-address space.

segment 🔒

user_ptr 🔒

Functions

init